Alias for public build, throw on resolve issue, better message
