+++ lib/internals.js
@@ -945,6 +945,7 @@ module.exports = (self) => {
       const vite = await import('vite');
       // The base public config
       const config = await vitePublicConfig({
+        sourceRoot: self.buildRootSource,
         input: self.getBuildInputs('public')
       });
       const postcssConfig = await self.getPostcssConfig(self.buildOptions, 'public');
