+++ lib/vite-apos-config.js
@@ -2,7 +2,7 @@ module.exports = async ({
   sourceRoot, input
 }) => {
   const vue = await import('@vitejs/plugin-vue');
-  const apos = await import('./vite-plugin-apostrophe-apos.mjs');
+  const apos = await import('./vite-plugin-apostrophe-alias.mjs');
 
   /** @type {import('vite').UserConfig} */
   return {
@@ -21,7 +21,10 @@ module.exports = async ({
       }
     },
     plugins: [
-      apos.default({ sourceRoot }),
+      apos.default({
+        id: 'apos',
+        sourceRoot
+      }),
       vue.default()
     ],
     build: {
