+++ lib/vite-public-config.js
@@ -1,9 +1,16 @@
 module.exports = async ({
-  input
+  sourceRoot, input
 }) => {
+  const apos = await import('./vite-plugin-apostrophe-alias.mjs');
 
   /** @type {import('vite').UserConfig} */
   return {
+    plugins: [
+      apos.default({
+        id: 'src',
+        sourceRoot
+      })
+    ],
     build: {
       rollupOptions: {
         input
